+++ CHANGELOG.md
@@ -20,6 +20,7 @@ v0.16.0 (xx dec 2018)
 - Fix loading of Sharepoint document @Garrcomm #1498
 - RTF writer: Round getPageSizeW and getPageSizeH to avoid decimals @Patrick64 #1493
 - Fix parsing of Office 365 documents @Timanx #1485
+- For RTF writers, sizes should should never have decimals @Samuel-BF #1536
 
 v0.15.0 (14 Jul 2018)
 ----------------------