+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-v0.16.0 (xx dec 2018)
+v0.16.0 (30 dec 2018)
 ----------------------
 ### Added
 - Add getVariableCount method in TemplateProcessor. @nicoder #1272